log flush 주기가 음수로 계산되어 실제로는 주기와 달리 바로바로 flush되던 버그수정.

+++ server/src/main/java/org/fastcatsearch/util/LimitTimeSizeLogger.java
@@ -28,7 +28,7 @@ public class LimitTimeSizeLogger {
 	public LimitTimeSizeLogger(File dir, String encoding, int flushPeriodInSeconds) {
 		this.bufferSize = 10000;
 		this.encoding = encoding;
-        this.flushPeriodInNanoseconds = flushPeriodInSeconds * 1000 * 1000 * 1000;
+        this.flushPeriodInNanoseconds = ((long) flushPeriodInSeconds) * 1000 * 1000 * 1000;
 		this.memoryData = newMemoryData();
         flushTimer = new Timer();
         long flushCheckPeriod = 500;
